app/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "celery>=5.5.3",
    "psycopg[binary]>=3.2.10",
    "pytest>=8.4.2",
    "pytest-mock>=3.15.1",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = [
    "ruff>=0.13.3",
]

[tool.ruff]
line-length = 100

target-version = "py312"

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "alembic",
    "alembic.ini"
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501", "E712", "F821", "UP017"]

fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"
